Initial version to get basic time logging stats
Class kimai that stores important parameters from time logging as
attributes.
Method stats that prints most important stats from attributes.
Class kimai is constructed from directory and file name of the kimai
exported times as csv, the year, and the number of vacation days. If not
given the file is assumed to be ../data/export.csv and the current year
and 0 vacation days are used.