Only consider vacation days in the current time frame
When vacation days are read from an input file, only consider those days
that are within the current Kimai time span to allow adding future and
past vacation documentation to that file.
Add parameter 'restrict_period' to workdays. If set to True, only
periods or parts of periods within self.__period are considered.
In vacation_days, restrict the the calculation of vacation spans to the
current Kimai period by setting restrict_period in the workdays call to
True. Check, whether a single vacation day is within the Kimai period
before increasing the counter.
Remove test print statement in init.
Closes #2.