Introduce working days in output
Rewrite method __format_timedelta to consider working days (= 8h).
Now method __format_timedelta shows 3 days for every day in a timedelta
and an additional day for every 8 hours in the time of timedelta. The
time must be smaller than 8 hours.
In method stats call __format_timedelta on every timedelta not just the
balance.
Rename attribute workingdays to workdays and method workdays to
work_days to be more consistent in the name scheme within Kimai class.
Rename __vaction to __vacdays and vacation to vacationdays to prepare
the attribute vacation storing the detailed vacation history.
Closes #5.